requirements.txt
django==4.2.3
environs==9.5.0
flake8==6.0.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-filter==23.2
 djangorestframework-simplejwt==5.2.2
